Guard against non-loaded sqlite driver

+++ lib/biggles/schema_helper.rb
@@ -41,7 +41,8 @@ module Biggles
     c.drop_table :biggles_scheduled
     c.drop_table :biggles_recurring
     c.drop_table :biggles_heartbeat
-    if c.is_a? ActiveRecord::ConnectionAdapters::SQLite3Adapter
+    if defined?(ActiveRecord::ConnectionAdapters::SQLite3Adapter) &&
+       c.is_a?(ActiveRecord::ConnectionAdapters::SQLite3Adapter)
       c.execute('DELETE FROM sqlite_sequence WHERE name = \'biggles_one_shot\'')
     else
       c.execute('DROP SEQUENCE biggles_one_shot_sequence')
